Fix SelectExpandBinder to take into account constant parameterization when using ModelId as part of a query(#1541)
When ModelId is used as part of a query in the SelectExpandBinder, it should respect the ODataQuerySettings.EnableConstantParameterization property. Otherwise, there is no way to force a generation of a SQL parameter in the resulting query. Which can result in polluting the SQL Server query plan cache with unwanted query plans (containing constants).